fix(#901): backendHttp.ts 추출 — positionUpload·boardingLockSync dup 제거
두 파일이 동일하게 정의한 getBackendUrl/fetchWithTimeout/REQUEST_TIMEOUT_MS를 한 모듈로.
Sonar 'new_duplicated_lines' 핵심 감점 제거. 3556 tests 유지.
